Refactor: 프로젝트 폴더 및 파일 구조 재정립
- 전체적으로 modules, views, control 폴더를 재구성하였습니다.
- modules 폴더에 데이터와 비즈니스 로직을 중점적으로 다루는 클래스를 재배치하였습니다.
- views 폴더에 사용자에게 힌트를 출력하는 로직을 다루는 클래스를 재배치하였습니다.
- controllers 폴더에 사용자의 입력 처리를 다루고 볼,스크라이크 계산을 하는 로직을 다루는 클래스를 재배치하였습니다.
