+++ src/views/HintView.js
